+++ ThirdPartyAdapters/dtexchange/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## DT Exchange Android Mediation Adapter Changelog
 
+#### Next Version
+- FyberMediationAdapter now extends from RtbAdapter
+
 #### Version 8.3.6.1
 - Updated the minimum required Android API level to 23.
 - Updated the minimum required Google Mobile Ads SDK version to 24.0.0.